+++ fatiando/gravmag/_prism.pyx
@@ -15,40 +15,59 @@ from cython.parallel cimport prange, parallel
 DTYPE = numpy.float
 ctypedef numpy.float_t DTYPE_T
 
+cdef inline double safe_atan2(double y, double x) nogil:
+    cdef double res
+    if y == 0:
+        res = 0
+    elif (y > 0) and (x < 0):
+        res = atan2(y, x) - 3.1415926535897931159979634685441851615906
+    elif (y < 0) and (x < 0):
+        res = atan2(y, x) + 3.1415926535897931159979634685441851615906
+    else:
+        res = atan2(y, x)
+    return res
+
+cdef inline double safe_log(double x) nogil:
+    cdef double res
+    if x == 0:
+        res = 0
+    else:
+        res = log(x)
+    return res
 
 cdef inline double kernelpot(double x, double y, double z, double r) nogil:
-    return (x*y*log(z + r) + y*z*log(x + r) + x*z*log(y + r)
-            - 0.5*x**2*atan2(z*y, x*r) - 0.5*y**2*atan2(z*x, y*r)
-            - 0.5*z**2*atan2(x*y, z*r))
+    return (x*y*safe_log(z + r) + y*z*safe_log(x + r) + x*z*safe_log(y + r)
+            - 0.5*x**2*safe_atan2(z*y, x*r) - 0.5*y**2*safe_atan2(z*x, y*r)
+            - 0.5*z**2*safe_atan2(x*y, z*r))
 
 # Minus in gravity because Nagy et al (2000) give the formula for the gradient
 # of the potential. Gravity is -grad(V).
 cdef inline double kernelx(double x, double y, double z, double r) nogil:
-    return -(y*log(z + r) + z*log(y + r) - x*atan2(z*y, x*r))
+    return -(y*safe_log(z + r) + z*safe_log(y + r) - x*safe_atan2(z*y, x*r))
 
 cdef inline double kernely(double x, double y, double z, double r) nogil:
-    return -(z*log(x + r) + x*log(z + r) - y*atan2(x*z, y*r))
+    return -(z*safe_log(x + r) + x*safe_log(z + r) - y*safe_atan2(x*z, y*r))
 
 cdef inline double kernelz(double x, double y, double z, double r) nogil:
-    return -(x*log(y + r) + y*log(x + r) - z*atan2(x*y, z*r))
+    return -(x*safe_log(y + r) + y*safe_log(x + r) - z*safe_atan2(x*y, z*r))
 
 cdef inline double kernelxx(double x, double y, double z, double r) nogil:
-    return -atan2(z*y, x*r)
+    return -safe_atan2(z*y, x*r)
 
 cdef inline double kernelxy(double x, double y, double z, double r) nogil:
-    return log(z + r)
+    return safe_log(z + r)
 
 cdef inline double kernelxz(double x, double y, double z, double r) nogil:
-    return log(y + r)
+    return safe_log(y + r)
 
 cdef inline double kernelyy(double x, double y, double z, double r) nogil:
-    return -atan2(z*x, y*r)
+    return -safe_atan2(z*x, y*r)
 
 cdef inline double kernelyz(double x, double y, double z, double r) nogil:
-    return log(x + r)
+    return safe_log(x + r)
 
 cdef inline double kernelzz(double x, double y, double z, double r) nogil:
-    return -atan2(x*y, z*r)
+    return -safe_atan2(x*y, z*r)
 
 @cython.wraparound(False)
 @cython.boundscheck(False)
@@ -302,7 +321,7 @@ def gxy(numpy.ndarray[DTYPE_T, ndim=1] xp not None,
         numpy.ndarray[DTYPE_T, ndim=1] res not None):
     cdef unsigned int l, size, i, j, k
     cdef numpy.ndarray[DTYPE_T, ndim=1] x, y, z
-    cdef DTYPE_T kernel, r, dx, dy, dz
+    cdef DTYPE_T kernel, r, dx, dy, dz, tmp1, tmp2
     size = len(xp)
     x = numpy.array([x2, x1], dtype=DTYPE)
     y = numpy.array([y2, y1], dtype=DTYPE)
@@ -316,7 +335,12 @@ def gxy(numpy.ndarray[DTYPE_T, ndim=1] xp not None,
                     dy = y[j] - yp[l]
                     for i in range(2):
                         dx = x[i] - xp[l]
-                        r = sqrt(dx**2 + dy**2 + dz**2)
+                        if dx == 0 and dy == 0 and dz < 0:
+                            tmp1 = 0.00001*(x2 - x1)
+                            tmp2 = 0.00001*(y2 - y1)
+                            r = sqrt(tmp1**2 + tmp2**2 + dz**2)
+                        else:
+                            r = sqrt(dx**2 + dy**2 + dz**2)
                         kernel = kernelxy(dx, dy, dz, r)
                         res[l] += ((-1.)**(i + j + k))*kernel*density
 
@@ -330,7 +354,7 @@ def gxz(numpy.ndarray[DTYPE_T, ndim=1] xp not None,
         numpy.ndarray[DTYPE_T, ndim=1] res not None):
     cdef unsigned int l, size, i, j, k
     cdef numpy.ndarray[DTYPE_T, ndim=1] x, y, z
-    cdef DTYPE_T kernel, r, dx, dy, dz
+    cdef DTYPE_T kernel, r, dx, dy, dz, tmp1, tmp2
     size = len(xp)
     x = numpy.array([x2, x1], dtype=DTYPE)
     y = numpy.array([y2, y1], dtype=DTYPE)
@@ -344,7 +368,12 @@ def gxz(numpy.ndarray[DTYPE_T, ndim=1] xp not None,
                     dy = y[j] - yp[l]
                     for i in range(2):
                         dx = x[i] - xp[l]
-                        r = sqrt(dx**2 + dy**2 + dz**2)
+                        if dx == 0 and dz == 0 and dy < 0:
+                            tmp1 = 0.00001*(x2 - x1)
+                            tmp2 = 0.00001*(z2 - z1)
+                            r = sqrt(tmp1**2 + tmp2**2 + dy**2)
+                        else:
+                            r = sqrt(dx**2 + dy**2 + dz**2)
                         kernel = kernelxz(dx, dy, dz, r)
                         res[l] += ((-1.)**(i + j + k))*kernel*density
 
@@ -386,7 +415,7 @@ def gyz(numpy.ndarray[DTYPE_T, ndim=1] xp not None,
         numpy.ndarray[DTYPE_T, ndim=1] res not None):
     cdef unsigned int l, size, i, j, k
     cdef numpy.ndarray[DTYPE_T, ndim=1] x, y, z
-    cdef DTYPE_T kernel, r, dx, dy, dz
+    cdef DTYPE_T kernel, r, dx, dy, dz, tmp1, tmp2
     size = len(xp)
     x = numpy.array([x2, x1], dtype=DTYPE)
     y = numpy.array([y2, y1], dtype=DTYPE)
@@ -400,7 +429,12 @@ def gyz(numpy.ndarray[DTYPE_T, ndim=1] xp not None,
                     dy = y[j] - yp[l]
                     for i in range(2):
                         dx = x[i] - xp[l]
-                        r = sqrt(dx**2 + dy**2 + dz**2)
+                        if dy == 0 and dz == 0 and dx < 0:
+                            tmp1 = 0.00001*(y2 - y1)
+                            tmp2 = 0.00001*(z2 - z1)
+                            r = sqrt(tmp1**2 + tmp2**2 + dx**2)
+                        else:
+                            r = sqrt(dx**2 + dy**2 + dz**2)
                         kernel = kernelyz(dx, dy, dz, r)
                         res[l] += ((-1.)**(i + j + k))*kernel*density