+++ fatiando/gravmag/prism.py
@@ -24,6 +24,15 @@ The gravitational fields are calculated using the formula of Nagy et al.
 * :func:`~fatiando.gravmag.prism.gyz`
 * :func:`~fatiando.gravmag.prism.gzz`
 
+.. warning::
+
+    The gxy, gxz, and gyz components have singularities when the computation
+    point is aligned with the corners of the prism on the bottom, east, and
+    north sides, respectively. In these cases, the above functions will move
+    the computation point slightly to avoid these singularities. Unfortunately,
+    this means that the result will not be as accurate **on those points**.
+
+
 **Magnetic**
 
 Available fields are the total-field anomaly (using the formula of
@@ -115,6 +124,8 @@ def potential(xp, yp, zp, prisms, dens=None):
         The field calculated on xp, yp, zp
 
     """
+    if xp.shape != yp.shape or xp.shape != zp.shape:
+        raise ValueError("Input arrays xp, yp, and zp must have same length!")
     size = len(xp)
     res = numpy.zeros(size, dtype=numpy.float)
     for prism in prisms:
@@ -161,6 +172,8 @@ def gx(xp, yp, zp, prisms, dens=None):
         The field calculated on xp, yp, zp
 
     """
+    if xp.shape != yp.shape or xp.shape != zp.shape:
+        raise ValueError("Input arrays xp, yp, and zp must have same length!")
     size = len(xp)
     res = numpy.zeros(size, dtype=numpy.float)
     for prism in prisms:
@@ -207,6 +220,8 @@ def gy(xp, yp, zp, prisms, dens=None):
         The field calculated on xp, yp, zp
 
     """
+    if xp.shape != yp.shape or xp.shape != zp.shape:
+        raise ValueError("Input arrays xp, yp, and zp must have same length!")
     size = len(xp)
     res = numpy.zeros(size, dtype=numpy.float)
     for prism in prisms:
@@ -253,6 +268,8 @@ def gz(xp, yp, zp, prisms, dens=None):
         The field calculated on xp, yp, zp
 
     """
+    if xp.shape != yp.shape or xp.shape != zp.shape:
+        raise ValueError("Input arrays xp, yp, and zp must have same length!")
     size = len(xp)
     res = numpy.zeros(size, dtype=numpy.float)
     for prism in prisms:
@@ -299,6 +316,8 @@ def gxx(xp, yp, zp, prisms, dens=None):
         The field calculated on xp, yp, zp
 
     """
+    if xp.shape != yp.shape or xp.shape != zp.shape:
+        raise ValueError("Input arrays xp, yp, and zp must have same length!")
     size = len(xp)
     res = numpy.zeros(size, dtype=numpy.float)
     for prism in prisms:
@@ -325,6 +344,14 @@ def gxy(xp, yp, zp, prisms, dens=None):
 
     .. note:: All input values in **SI** units(!) and output in **Eotvos**!
 
+    .. warning::
+
+        This component has singularities when the computation
+        point is aligned with the corners of the prism on the bottom side.
+        In these cases, the computation point slightly to avoid these
+        singularities. Unfortunately, this means that the result will not be as
+        accurate **on those points**.
+
     Parameters:
 
     * xp, yp, zp : arrays
@@ -345,6 +372,8 @@ def gxy(xp, yp, zp, prisms, dens=None):
         The field calculated on xp, yp, zp
 
     """
+    if xp.shape != yp.shape or xp.shape != zp.shape:
+        raise ValueError("Input arrays xp, yp, and zp must have same length!")
     size = len(xp)
     res = numpy.zeros(size, dtype=numpy.float)
     for prism in prisms:
@@ -371,6 +400,14 @@ def gxz(xp, yp, zp, prisms, dens=None):
 
     .. note:: All input values in **SI** units(!) and output in **Eotvos**!
 
+    .. warning::
+
+        This component has singularities when the computation
+        point is aligned with the corners of the prism on the east side.
+        In these cases, the computation point slightly to avoid these
+        singularities. Unfortunately, this means that the result will not be as
+        accurate **on those points**.
+
     Parameters:
 
     * xp, yp, zp : arrays
@@ -391,6 +428,8 @@ def gxz(xp, yp, zp, prisms, dens=None):
         The field calculated on xp, yp, zp
 
     """
+    if xp.shape != yp.shape or xp.shape != zp.shape:
+        raise ValueError("Input arrays xp, yp, and zp must have same length!")
     size = len(xp)
     res = numpy.zeros(size, dtype=numpy.float)
     for prism in prisms:
@@ -437,6 +476,8 @@ def gyy(xp, yp, zp, prisms, dens=None):
         The field calculated on xp, yp, zp
 
     """
+    if xp.shape != yp.shape or xp.shape != zp.shape:
+        raise ValueError("Input arrays xp, yp, and zp must have same length!")
     size = len(xp)
     res = numpy.zeros(size, dtype=numpy.float)
     for prism in prisms:
@@ -463,6 +504,14 @@ def gyz(xp, yp, zp, prisms, dens=None):
 
     .. note:: All input values in **SI** units(!) and output in **Eotvos**!
 
+    .. warning::
+
+        This component has singularities when the computation
+        point is aligned with the corners of the prism on the north side.
+        In these cases, the computation point slightly to avoid these
+        singularities. Unfortunately, this means that the result will not be as
+        accurate **on those points**.
+
     Parameters:
 
     * xp, yp, zp : arrays
@@ -483,6 +532,8 @@ def gyz(xp, yp, zp, prisms, dens=None):
         The field calculated on xp, yp, zp
 
     """
+    if xp.shape != yp.shape or xp.shape != zp.shape:
+        raise ValueError("Input arrays xp, yp, and zp must have same length!")
     size = len(xp)
     res = numpy.zeros(size, dtype=numpy.float)
     for prism in prisms:
@@ -529,6 +580,8 @@ def gzz(xp, yp, zp, prisms, dens=None):
         The field calculated on xp, yp, zp
 
     """
+    if xp.shape != yp.shape or xp.shape != zp.shape:
+        raise ValueError("Input arrays xp, yp, and zp must have same length!")
     size = len(xp)
     res = numpy.zeros(size, dtype=numpy.float)
     for prism in prisms:
@@ -578,7 +631,7 @@ def tf(xp, yp, zp, prisms, inc, dec, pmag=None):
         The field calculated on xp, yp, zp
 
     """
-    if len(xp) != len(yp) != len(zp):
+    if xp.shape != yp.shape or xp.shape != zp.shape:
         raise ValueError("Input arrays xp, yp, and zp must have same length!")
     size = len(xp)
     res = numpy.zeros(size, dtype=numpy.float)
@@ -635,7 +688,7 @@ def bx(xp, yp, zp, prisms, pmag=None):
         The x component of the magnetic induction
 
     """
-    if xp.shape != yp.shape != zp.shape:
+    if xp.shape != yp.shape or xp.shape != zp.shape:
         raise ValueError("Input arrays xp, yp, and zp must have same shape!")
     if pmag is not None:
         mx, my, mz = pmag
@@ -681,7 +734,7 @@ def by(xp, yp, zp, prisms, pmag=None):
         The y component of the magnetic induction
 
     """
-    if xp.shape != yp.shape != zp.shape:
+    if xp.shape != yp.shape or xp.shape != zp.shape:
         raise ValueError("Input arrays xp, yp, and zp must have same shape!")
     if pmag is not None:
         mx, my, mz = pmag
@@ -727,7 +780,7 @@ def bz(xp, yp, zp, prisms, pmag=None):
         The z component of the magnetic induction
 
     """
-    if xp.shape != yp.shape != zp.shape:
+    if xp.shape != yp.shape or xp.shape != zp.shape:
         raise ValueError("Input arrays xp, yp, and zp must have same shape!")
     if pmag is not None:
         mx, my, mz = pmag
@@ -772,7 +825,7 @@ def kernelxx(xp, yp, zp, prism):
         The effect calculated on the computation points.
 
     """
-    if xp.shape != yp.shape != zp.shape:
+    if xp.shape != yp.shape or xp.shape != zp.shape:
         raise ValueError("Input arrays xp, yp, and zp must have same shape!")
     res = numpy.zeros(len(xp), dtype=numpy.float)
     x1, x2 = prism.x1, prism.x2
@@ -807,7 +860,7 @@ def kernelyy(xp, yp, zp, prism):
         The effect calculated on the computation points.
 
     """
-    if xp.shape != yp.shape != zp.shape:
+    if xp.shape != yp.shape or xp.shape != zp.shape:
         raise ValueError("Input arrays xp, yp, and zp must have same shape!")
     res = numpy.zeros(len(xp), dtype=numpy.float)
     x1, x2 = prism.x1, prism.x2
@@ -842,7 +895,7 @@ def kernelzz(xp, yp, zp, prism):
         The effect calculated on the computation points.
 
     """
-    if xp.shape != yp.shape != zp.shape:
+    if xp.shape != yp.shape or xp.shape != zp.shape:
         raise ValueError("Input arrays xp, yp, and zp must have same shape!")
     res = numpy.zeros(len(xp), dtype=numpy.float)
     x1, x2 = prism.x1, prism.x2
@@ -877,7 +930,7 @@ def kernelxy(xp, yp, zp, prism):
         The effect calculated on the computation points.
 
     """
-    if xp.shape != yp.shape != zp.shape:
+    if xp.shape != yp.shape or xp.shape != zp.shape:
         raise ValueError("Input arrays xp, yp, and zp must have same shape!")
     res = numpy.zeros(len(xp), dtype=numpy.float)
     x1, x2 = prism.x1, prism.x2
@@ -912,7 +965,7 @@ def kernelxz(xp, yp, zp, prism):
         The effect calculated on the computation points.
 
     """
-    if xp.shape != yp.shape != zp.shape:
+    if xp.shape != yp.shape or xp.shape != zp.shape:
         raise ValueError("Input arrays xp, yp, and zp must have same shape!")
     res = numpy.zeros(len(xp), dtype=numpy.float)
     x1, x2 = prism.x1, prism.x2
@@ -947,7 +1000,7 @@ def kernelyz(xp, yp, zp, prism):
         The effect calculated on the computation points.
 
     """
-    if xp.shape != yp.shape != zp.shape:
+    if xp.shape != yp.shape or xp.shape != zp.shape:
         raise ValueError("Input arrays xp, yp, and zp must have same shape!")
     res = numpy.zeros(len(xp), dtype=numpy.float)
     x1, x2 = prism.x1, prism.x2